Desktop/voiceSignature/requirements.txt
torch>=1.9.0
torchaudio>=0.9.0
transformers>=4.20.0
speechbrain>=0.5.14
librosa>=0.9.0
numpy>=1.21.0
scipy>=1.7.0
soundfile>=0.10.0
resemblyzer>=0.1.1.dev0
nemo-toolkit[asr]>=1.15.0
omegaconf>=2.3.0
tqdm>=4.62.0
matplotlib>=3.5.0
seaborn>=0.11.0 